# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "FlowLink API",
    "description" : "Project and portfolio management with a resource-constrained scheduling optimizer.",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/tasks/{taskId}/comments/{commentId}" : {
      "put" : {
        "tags" : [ "comment-controller" ],
        "operationId" : "updateComment",
        "parameters" : [ {
          "name" : "taskId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "commentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "content" : {
                    "type" : "string",
                    "maxLength" : 10000,
                    "minLength" : 0
                  },
                  "attachments" : {
                    "type" : "array",
                    "items" : {
                      "type" : "string",
                      "format" : "binary"
                    }
                  }
                },
                "required" : [ "content" ]
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommentDTO"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "comment-controller" ],
        "operationId" : "deleteComment",
        "parameters" : [ {
          "name" : "taskId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "commentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/tasks/{taskId}/comments" : {
      "get" : {
        "tags" : [ "comment-controller" ],
        "operationId" : "getComments",
        "parameters" : [ {
          "name" : "taskId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PagedResponseCommentDTO"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "comment-controller" ],
        "operationId" : "addComment",
        "parameters" : [ {
          "name" : "taskId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "parentCommentId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "content" : {
                    "type" : "string",
                    "maxLength" : 10000,
                    "minLength" : 0
                  },
                  "attachments" : {
                    "type" : "array",
                    "items" : {
                      "type" : "string",
                      "format" : "binary"
                    }
                  }
                },
                "required" : [ "content" ]
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommentDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{taskId}/comments/{commentId}/reactions" : {
      "post" : {
        "tags" : [ "comment-controller" ],
        "operationId" : "reactToComment",
        "parameters" : [ {
          "name" : "taskId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "commentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "type",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "LIKE", "DISLIKE" ]
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommentDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/optimization/simulate" : {
      "post" : {
        "tags" : [ "optimization-controller" ],
        "operationId" : "simulate",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OptimizationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OptimizationResultDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/optimization/apply" : {
      "post" : {
        "tags" : [ "optimization-controller" ],
        "operationId" : "apply",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ApplyOptimizationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AppliedSchedule"
                }
              }
            }
          }
        }
      }
    },
    "/api/notifications/mark-as-read" : {
      "post" : {
        "tags" : [ "notification-controller" ],
        "operationId" : "markAsRead",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "items" : {
                  "type" : "integer",
                  "format" : "int32"
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/notifications" : {
      "get" : {
        "tags" : [ "notification-controller" ],
        "operationId" : "getNotifications",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PagedResponseNotificationDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "CommentDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "taskId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "authorId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "authorName" : {
            "type" : "string"
          },
          "authorProfilePicture" : {
            "type" : "string"
          },
          "content" : {
            "type" : "string"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          },
          "editedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "attachments" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "likeCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "dislikeCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "likedByCurrentUser" : {
            "type" : "boolean"
          },
          "dislikedByCurrentUser" : {
            "type" : "boolean"
          },
          "replies" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CommentDTO"
            }
          }
        }
      },
...
      "OptimizationRequest" : {
        "type" : "object",
        "properties" : {
          "projectKeys" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            },
            "maxItems" : 50,
            "minItems" : 0
          },
          "alpha" : {
            "type" : "number",
            "format" : "double",
            "maximum" : 1.0,
            "minimum" : 0.0
          },
          "beta" : {
            "type" : "number",
            "format" : "double",
            "maximum" : 1.0,
            "minimum" : 0.0
          },
          "horizonStart" : {
            "type" : "string",
            "format" : "date"
          }
        }
      },
      "OptimizationMetricsDTO" : {
        "type" : "object",
        "properties" : {
          "weightedTardiness" : {
            "type" : "number",
            "format" : "double"
          },
          "makespan" : {
            "type" : "integer",
            "format" : "int32"
          },
          "objectiveValue" : {
            "type" : "number",
            "format" : "double"
          },
          "totalTasks" : {
            "type" : "integer",
            "format" : "int32"
          },
          "tasksOnTime" : {
            "type" : "integer",
            "format" : "int32"
          },
          "tasksLate" : {
            "type" : "integer",
            "format" : "int32"
          },
          "resourceConflicts" : {
            "type" : "integer",
            "format" : "int32"
          },
          "feasible" : {
            "type" : "boolean"
          }
        }
      },
      "OptimizationResultDTO" : {
        "type" : "object",
        "properties" : {
          "suggestions" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TaskScheduleSuggestionDTO"
            }
          },
          "originalMetrics" : {
            "$ref" : "#/components/schemas/OptimizationMetricsDTO"
          },
          "optimizedMetrics" : {
            "$ref" : "#/components/schemas/OptimizationMetricsDTO"
          },
          "tasksShifted" : {
            "type" : "integer",
            "format" : "int32"
          },
          "chosenRule" : {
            "type" : "string"
          },
          "skippedTaskKeys" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "TaskScheduleSuggestionDTO" : {
        "type" : "object",
        "properties" : {
          "taskKey" : {
            "type" : "string"
          },
          "projectKey" : {
            "type" : "string"
          },
          "summary" : {
            "type" : "string"
          },
          "assignee" : {
            "type" : "string"
          },
          "originalStartDate" : {
            "type" : "string",
            "format" : "date"
          },
          "originalDueDate" : {
            "type" : "string",
            "format" : "date"
          },
          "suggestedStartDate" : {
            "type" : "string",
            "format" : "date"
          },
          "suggestedDueDate" : {
            "type" : "string",
            "format" : "date"
          },
          "priorityWeight" : {
            "type" : "integer",
            "format" : "int32"
          },
          "tardinessDays" : {
            "type" : "integer",
            "format" : "int32"
          },
          "isCritical" : {
            "type" : "boolean"
          },
          "wasShifted" : {
            "type" : "boolean"
          }
        }
      },
      "ApplyOptimizationRequest" : {
        "type" : "object",
        "properties" : {
          "projectKeys" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            },
            "maxItems" : 50,
            "minItems" : 0
          },
          "alpha" : {
            "type" : "number",
            "format" : "double",
            "maximum" : 1.0,
            "minimum" : 0.0
          },
          "beta" : {
            "type" : "number",
            "format" : "double",
            "maximum" : 1.0,
            "minimum" : 0.0
          },
          "horizonStart" : {
            "type" : "string",
            "format" : "date"
          },
          "acceptedTaskKeys" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            },
            "maxItems" : 5000,
            "minItems" : 0
          }
        }
      },
      "AppliedSchedule" : {
        "type" : "object",
        "properties" : {
          "tasksUpdated" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "PagedResponseCommentDTO" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CommentDTO"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "hasNext" : {
            "type" : "boolean"
          }
        }
      },
...
      "NotificationDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          },
          "isRead" : {
            "type" : "boolean"
          },
          "type" : {
            "type" : "string",
            "enum" : [ "PROJECT_INVITATION", "PROJECT_UPDATED", "MEMBER_REMOVED", "TASK_ASSIGNED", "TASK_UPDATED", "TASK_COMMENT", "TASK_DELETED", "COMMENT_REPLY", "COMMENT_REACTION" ]
          },
          "link" : {
            "type" : "string"
          }
        }
      },
      "PagedResponseNotificationDTO" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/NotificationDTO"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "hasNext" : {
            "type" : "boolean"
          }
        }
      }
    }
  }
}